+++ TODO
@@ -4,8 +4,8 @@ Project:
 
 Ideas:
     - 13/01/2022
-        - Hacer componentes grid (componentes de gestion de posicionamiento de la vista)
-        - La vista de leer posts que este centrada
+        - Hacer componentes de gestion del display de los componentes @done
+        - La vista de leer posts que este centrada @done
         - Meter un post de ejemplo para ver como queda
         - Meter redes de contacto en el header
         - Obtener metadata de un post
