--- posts/posts/test.md
@@ -1,7 +1,0 @@
----
-title: "Two Forms of Pre-rendering"
-date: "2020-01-01"
-tag: "tag"
----
-
-test
